modify training session controller

+++ app/Http/Controllers/Backend/TrainingSessionsController.php
@@ -3,11 +3,10 @@
 namespace App\Http\Controllers\Backend;
 
 use App\DataTables\TrainingSessionDataTable;
-use App\Http\Controllers\Controller;
 use App\Http\Requests\TrainingSessionRequest;
+use App\Models\Coach;
 use App\Models\Gym;
 use App\Models\TrainingSession;
-use Illuminate\Http\Request;
 
 class TrainingSessionsController extends CommonController
 {
@@ -22,22 +21,25 @@ class TrainingSessionsController extends CommonController
 
     public function create()
     {
-        $gyms=Gym::all()->pluck('name','id');
-       return view('dashboard.trainingSession.create',['gyms'=>$gyms]);
+        $gyms= Gym::all()->pluck('name','id');
+        $coaches = Coach::all()->pluck('name', 'id');
+        return view('dashboard.trainingSession.create', ['coaches' => $coaches,'gyms'=>$gyms]);
     }
 
 
     public function store(TrainingSessionRequest $request)
     {
-       TrainingSession::create($request->validated());
-       return to_route('dashboard.training-sessions.index');
+        $trainingSession= TrainingSession::create($request->validated());
+        $trainingSession->coaches()->attach($request->coach_id);
+        return to_route('dashboard.training-sessions.index');
     }
 
 
     public function edit(TrainingSession $trainingSession)
     {
-        $gyms=Gym::all()->pluck('name','id');
-       return view('dashboard.trainingSession.edit',['gyms'=>$gyms,'trainingSession'=>$trainingSession]);
+        $arr=TrainingSession::doesntHave('users')->get();
+        $gyms = Gym::all()->pluck('name', 'id');
+        return view('dashboard.trainingSession.edit', ['gyms' => $gyms, 'trainingSession' => $trainingSession,'arr'=>$arr]);
     }
 
 
@@ -46,11 +48,12 @@ class TrainingSessionsController extends CommonController
         $trainingSession->update($request->validated());
         return to_route('dashboard.training-sessions.index');
     }
-    
+
     public function destroy(TrainingSession $trainingSession)
     {
+        $arr=TrainingSession::doesntHave('users')->get();
+        if ($arr->contains('id',$trainingSession->id))
         $trainingSession->delete();
         return to_route('dashboard.training-sessions.index');
-        
     }
 }
